+++ lib/epic_failure.dart
@@ -3,10 +3,10 @@ library epic_failure;
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 
-export 'src/core/epic_sitter.dart';
+export 'src/core/failure_manager.dart';
 
 part 'src/functions/on_failure_thrown.dart';
 
-part 'src/models/epic_check.dart';
 part 'src/models/epic_failure.dart';
-part 'src/models/epic_print.dart';
+part 'src/models/failure_code.dart';
+part 'src/models/predetermined_failure.dart';