+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.0] - A useful release!
+
+- Functioning `EpicHandler` to register predetermined failures and generate `EpicFailure`s.
+- README.md with example code for use.
+
 ## [0.0.1] - TODO: Add release date.
 
-* TODO: Describe initial release.
+- TODO: Describe initial release.